requirements.txt
Flask
gunicorn
requests
flask_sqlalchemy
psycopg2-binary
